Rework
Ditched the React app in favor of a Next app.

Routing has been done.

Need to look out for .env file data, and data as a whole. The requests were taken from a previous next project but are correct. We only need to adapt them to the back schemas once the back will be done.